From API to DB working

+++ api/src/models/Country.js
@@ -1,6 +1,4 @@
 const { DataTypes, Op } = require('sequelize');
-// Exportamos una funcion que define el modelo
-// Luego le injectamos la conexion a sequelize.
 
 module.exports = (sequelize) => {
   const Pais = sequelize.define('pais', {
@@ -10,7 +8,7 @@ module.exports = (sequelize) => {
     continente: { type: DataTypes.STRING, allowNull: false },
     capital: { type: DataTypes.STRING, allowNull: false },
     subregion: { type: DataTypes.STRING },
-    area: { type: DataTypes.INTEGER },
+    area: { type: DataTypes.DOUBLE },
     poblacion: { type: DataTypes.INTEGER }
   });
